pushing this for when i go to banff tomorrow. Guess I wont be debugging :^)

+++ stage-8/splsh.c
@@ -2,13 +2,15 @@
 #include "../stage-2/RASTER.H"
 #include "../stage-2/objects.h"
 #include "CONST8.H"
-
+#include "splash.h"
+#include <stdio.h>
+#include <osbind.h>
 
 
 int main()
 {
     
-    int *base = Physbase();
+    UINT8 *base = Physbase();
 
     Cnecin();
     clearScreen(base);
